refactored CardWeddingInfo
