--- src/views/components/organisms/CardWeddingInfo/InfoBasic.tsx
@@ -1,17 +1,0 @@
-const InfoBasic = () => {
-  return (
-    <section className="FlexCenter flex-col gap-4">
-      <h5 className="text-Green-dark">
-        SATURDAY JUNE 18, 2022 at 10:00 in the morning
-      </h5>
-      <h5 className="text-center text-Green-dark">
-        WOODLAND FOREST INN
-        <br />
-        12458 redwood drive anywhere city, CA
-      </h5>
-      <p className="text-Green-dark">reception to follow</p>
-    </section>
-  );
-};
-
-export default InfoBasic;
