--- src/views/components/organisms/CardWeddingInfo/InfoDetail.tsx
@@ -1,14 +1,0 @@
-const InfoDetail = () => {
-  return (
-    <section className=" flex justify-center">
-      <p className="w-4/5 flex text-Green-dark leading-5">
-        Lorem, ipsum dolor sit amet consectetur adipisicing elit. Ratione,
-        deleniti dignissimos! Laudantium ex officiis voluptate dolor, atque
-        debitis fugiat dignissimos nesciunt dicta nostrum ab ut aliquid error
-        amet laboriosam iste!
-      </p>
-    </section>
-  );
-};
-
-export default InfoDetail;
